消除了一些警告

+++ MSNumberScrollAnimatedView/MSNumberScrollAnimatedView.m
@@ -124,8 +124,8 @@
     NSInteger number = [numberText integerValue];
     NSMutableArray *scrollNumbers = [NSMutableArray array];
     // 添加要滚动的数字
-    for (NSUInteger i = 0; i < self.density + 1; i++) {
-        [scrollNumbers addObject:[NSString stringWithFormat:@"%lu", (number+i) % 10]];
+    for (NSInteger i = 0; i < self.density + 1; i++) {
+        [scrollNumbers addObject:[NSString stringWithFormat:@"%u", (unsigned int)((number+i) % 10)]];
     }
     [scrollNumbers addObject:numberText];
     // 创建 scrollLayer 的内容，数字降序排序
